ploopyco/opt_encoder: minimize public API
This commit defines stricter typing and remove unnecessary symbols from
the header.
